renamed package names as suggested
