+++ web-sample-app/src/test/java/com/tpg/quality/web/sample/or/HomePageObjects.java
@@ -2,9 +2,10 @@ package com.tpg.quality.web.sample.or;
 
 import org.openqa.selenium.WebDriver;
 import org.springframework.stereotype.Component;
-import com.tpg.quality.web.utility.custom_annotation.FindBy;
-import com.tpg.quality.web.utility.custom_pagefactory.CustomPageFactory;
-import com.tpg.quality.web.utility.custom_webelements.Button;
+
+import com.tpg.quality.web.annotation.FindBy;
+import com.tpg.quality.web.pagefactory.CustomPageFactory;
+import com.tpg.quality.web.webelements.Button;
 
 @Component
 public class HomePageObjects {
