+++ web/src/main/java/com/tpg/quality/web/driver/CustomFirefox.java
@@ -1,19 +1,20 @@
-package com.tpg.quality.web.utility.custom_driver;
+package com.tpg.quality.web.driver;
 
 import java.lang.reflect.Proxy;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.ie.InternetExplorerDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.springframework.stereotype.Component;
 import com.tpg.quality.web.utility.MyInvocationHandler;
 
 @Component
-public class CustomIE {
+public class CustomFirefox {
+
 	String location = System.getProperty("user.dir");
 	private WebDriver driver;
 
-	public WebDriver getIEDriver() {
-		System.setProperty("webdriver.ie.driver", location + "\\resources\\iedriver.exe");
-		driver = new InternetExplorerDriver();
+	public WebDriver getFirefoxDriver() {
+		System.setProperty("webdriver.gecko.driver", location + "\\resources\\geckodriver.exe");
+		driver = new FirefoxDriver();
 		driver = (WebDriver) Proxy.newProxyInstance(WebDriver.class.getClassLoader(), new Class[] { WebDriver.class },
 				new MyInvocationHandler(driver));
 		return driver;
