+++ web/src/main/java/com/tpg/quality/web/driver/CustomIE.java
@@ -1,20 +1,19 @@
-package com.tpg.quality.web.utility.custom_driver;
+package com.tpg.quality.web.driver;
 
 import java.lang.reflect.Proxy;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.ie.InternetExplorerDriver;
 import org.springframework.stereotype.Component;
 import com.tpg.quality.web.utility.MyInvocationHandler;
 
 @Component
-public class CustomFirefox {
-
+public class CustomIE {
 	String location = System.getProperty("user.dir");
 	private WebDriver driver;
 
-	public WebDriver getFirefoxDriver() {
-		System.setProperty("webdriver.gecko.driver", location + "\\resources\\geckodriver.exe");
-		driver = new FirefoxDriver();
+	public WebDriver getIEDriver() {
+		System.setProperty("webdriver.ie.driver", location + "\\resources\\iedriver.exe");
+		driver = new InternetExplorerDriver();
 		driver = (WebDriver) Proxy.newProxyInstance(WebDriver.class.getClassLoader(), new Class[] { WebDriver.class },
 				new MyInvocationHandler(driver));
 		return driver;
