+++ web/src/main/java/com/tpg/quality/web/pagefactory/CustomPageFactory.java
@@ -1,11 +1,11 @@
-package com.tpg.quality.web.utility.custom_pagefactory;
+package com.tpg.quality.web.pagefactory;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
 
-import com.tpg.quality.web.utility.custom_annotation.CustomElementLocatorFactory;
+import com.tpg.quality.web.annotation.CustomElementLocatorFactory;
 
 public class CustomPageFactory extends PageFactory {
